use AC_PROG_CC_C99 + print fix (https://lists.gforge.inria.fr/pipermail/ecm-discuss/2020-April/004485.html)
git-svn-id: svn://scm.gforge.inria.fr/svnroot/ecm/trunk@3081 404564d9-a503-0410-82bf-e18ce2cf3989

+++ main.c
@@ -1375,10 +1375,10 @@ main (int argc, char *argv[])
               exit (EXIT_FAILURE);
             }
           else if (verbose >= OUTPUT_VERBOSE)
-              fprintf (stdout, "Reading batch product (of %"PRIu64" bits) of "
+              fprintf (stdout, "Reading batch product (of %lu bits) of "
                                "primes up to B1=%1.0f from %s took %ldms\n", 
-                               mpz_sizeinbase (params->batch_s, 2), B1,
-                               loadfile_s, cputime () - st);
+                       (unsigned long) mpz_sizeinbase (params->batch_s, 2), B1,
+                       loadfile_s, cputime () - st);
         }
 
       /* set parameters that may change from one curve to another */
